Arbitrable: Improve multiple versions handling
